fst-04-30-2025 (#9)
* fst-lcc-04-23-2025

Updated the names of our tools in license and copyright notices to reflect changes we have made. Also, the tools name will be changed in the DOI, so need to do this.

* lcc-04-23-2025

The name of the LCAC tool has been changed to reflect new changes. Everything is LCC now.

* namechange

* Delete apps/.DS_Store

* Delete .DS_Store

* Delete apps/lcc/.Rhistory

* Delete apps/lcc/.DS_Store

* fst-04-30-2025

Added Energy Costs Sankey
Added multiple checkbox for product intensities
Added MWh option for product intensity
Added inputs memory for product intensity

* fst-04-30-2025-02

Added some language in the front page of the tool

---------

Co-authored-by: Ovais Ahmed Khan <ovais.ahmed211@gmail.com>
Co-authored-by: OvaisAK <142067347+OvaisAK@users.noreply.github.com>
Co-authored-by: Biju Jacob <github@biju.com>