Updated the input sheet logic and cell references.
Updated the input sheet to base its logic on the primary, secondary, and switched fuel selection. Added new columns to accurately aggregate energy, cost and emission savings based on this logic.

Corrected table array references in the code.

Added if required, install instructions for the application's dependencies before loading to make it more accessible to users without prior R experience to run it locally.

Corrected the decarbonization measure category call to how it is in the input sheet.